requirements.txt
openai>=0.28.0
pylint>=3.0.0
click>=7.1.2
pyyaml>=6.0.0
PyGithub>=2.1.1
rich>=13.0.0
anthropic>=0.3.0  # For Claude models
deepseek>=0.1.0   # For DeepSeek models
